requirements.txt
# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6  # 文件上传支持
jinja2==3.1.3  # 模板引擎

# 数据库
sqlalchemy==2.0.25

# HTTP客户端
httpx==0.26.0

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0

# 图片处理
Pillow==10.2.0

# 工具库
python-dotenv==1.0.0  # 环境变量管理
